Refactor Date Widget into seperate component on Ascent Modal
