+++ client/src/components/peakLists/import/index.tsx
@@ -29,7 +29,7 @@ import {
   ADD_MOUNTAIN_COMPLETION,
   MountainCompletionSuccessResponse,
   MountainCompletionVariables,
-} from '../detail/completionModal/MountainCompletionModal';
+} from '../detail/completionModal/queries';
 import {
   horizontalPadding,
 } from '../detail/MountainRow';
